+++ nodeconductor/server/admin/menu.py
@@ -1,29 +1,7 @@
-import hashlib
-
-from admin_tools import utils
 from admin_tools.menu import items, Menu
-from django.core.cache import cache
 from django.core.urlresolvers import reverse
 from django.utils.text import capfirst
 from django.utils.translation import ugettext_lazy as _
-
-
-patched_filter_models = utils.filter_models
-
-
-def _filter_models(request, models, exclude):
-    key = request.user.uuid.hex + ''.join(models) + ''.join(exclude)
-    hashed_key = hashlib.sha256(key).hexdigest()
-    if hashed_key in cache:
-        result = cache.get(hashed_key)
-    else:
-        result = patched_filter_models(request, models, exclude)
-        cache.set(hashed_key, result, 60 * 60)
-
-    return result
-
-# This patch is required to decrease number of DB queries
-utils.filter_models = _filter_models
 
 
 class CustomAppList(items.AppList):
@@ -62,11 +40,21 @@ class CustomMenu(Menu):
 
     IAAS_CLOUDS = (
         'nodeconductor_assembly_waldur.packages.*',
-        'nodeconductor_assembly_waldur.invoices.*',
         'nodeconductor_azure.*',
         'nodeconductor_openstack.*',
-        'nodeconductor_plus.aws.*',
-        'nodeconductor_plus.digitalocean.*',
+        'nodeconductor_aws.*',
+        'nodeconductor_digitalocean.*',
+    )
+
+    USERS = (
+        'nodeconductor.core.models.User',
+        'nodeconductor.core.models.SSHPublicKey',
+        'nodeconductor.users.models.Invitation',
+    )
+
+    ACCOUNTING = (
+        'nodeconductor_assembly_waldur.invoices.*',
+        'nodeconductor.cost_tracking.*',
     )
 
     APPLICATION_PROVIDERS = (
@@ -80,32 +68,30 @@ class CustomMenu(Menu):
     )
 
     SUPPORT_MODULES = (
-        'nodeconductor_plus.plans.*',
-        'nodeconductor_plus.premium_support.*',
+        'nodeconductor_assembly_waldur.support.*',
     )
 
     def __init__(self, **kwargs):
         Menu.__init__(self, **kwargs)
         self.children += [
             items.MenuItem(_('Dashboard'), reverse('admin:index')),
-            CustomAppList(
-                _('Core'),
-                exclude=('django.core.*',
-                         'rest_framework.authtoken.*',
-                         'nodeconductor.core.*',
-                         'nodeconductor.structure.*',
-                         ) + self.IAAS_CLOUDS + self.APPLICATION_PROVIDERS + self.SUPPORT_MODULES
+            items.ModelList(
+                _('Users'),
+                models=self.USERS
             ),
             items.ModelList(
                 _('Structure'),
-                models=('nodeconductor.core.*',
-                        'nodeconductor_organization.*',
-                        'nodeconductor.structure.*',
-                        'nodeconductor_assembly_itacloud.template.*',  # Hack to show template groups in admin
-                        )
+                models=(
+                    'nodeconductor.structure.*',
+                )
             ),
             CustomAppList(
-                _('IaaS clouds'),
+                _('Accounting'),
+                models=self.ACCOUNTING,
+            ),
+
+            CustomAppList(
+                _('Providers'),
                 models=self.IAAS_CLOUDS,
             ),
             CustomAppList(
@@ -113,7 +99,22 @@ class CustomMenu(Menu):
                 models=self.APPLICATION_PROVIDERS,
             ),
             CustomAppList(
-                _('Subscriptions and support'),
+                _('Support'),
                 models=self.SUPPORT_MODULES,
             ),
+            CustomAppList(
+                _('Utilities'),
+                exclude=('django.core.*',
+                         'django_openid_auth.*',
+                         'rest_framework.authtoken.*',
+                         'nodeconductor.core.*',
+                         'nodeconductor.structure.*',
+                         )
+                        + self.IAAS_CLOUDS
+                        + self.APPLICATION_PROVIDERS
+                        + self.SUPPORT_MODULES
+                        + self.ACCOUNTING
+                        + self.USERS
+            ),
+
         ]